Add sm-table-widget.
This module is transplanted from the `editor` repo, where it was called
"my-table-widget".  The test module `sm-table-widget-test` configures
the table the same way that `editor/open-files-dialog.cc` does.

The code in this commit should behave exactly like it did when it was
in the `editor` repo.

My intention now is to make some improvements to the interface and to
the column resizing behavior.